Refactor Issue related functionalities to separate file
The commit introduces a new source file `issue.go` which now hosts the Issue related functionalities and types which were earlier a part of `main.go`. Corresponding test cases were also re-routed from `main_test.go` to a newly created test file `issue_test.go`. This segregation improves code readability and maintainability.
